feat(api/user): Move user api to admin section
